requirements.txt
pyrogram
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
pydantic
alembic
tgcrypto